+++ agent/config/config.go
@@ -93,6 +93,14 @@ const (
 
 	// pauseContainerTarball is the path to the pause container tarball
 	pauseContainerTarballPath = "/images/amazon-ecs-pause.tar"
+
+	// DefaultTaskMetadataSteadyStateRate is set as 40. This is arrived from our benchmarking
+	// results where task endpoint can handle 4000 rps effectively. Here, 100 containers
+	// will be able to send out 40 rps.
+	DefaultTaskMetadataSteadyStateRate = 40
+
+	// DefaultTaskMetadataBurstRate is set to handle 60 burst requests at once
+	DefaultTaskMetadataBurstRate = 60
 )
 
 var (
@@ -346,6 +354,7 @@ func environmentConfig() (Config, error) {
 	}
 	containerMetadataEnabled := utils.ParseBool(os.Getenv("ECS_ENABLE_CONTAINER_METADATA"), false)
 	dataDirOnHost := os.Getenv("ECS_HOST_DATA_DIR")
+	steadyStateRate, burstRate := getTaskMetadataThrottles()
 
 	if len(errs) > 0 {
 		err = utils.NewMultiError(errs...)
@@ -391,6 +400,8 @@ func environmentConfig() (Config, error) {
 		DataDirOnHost:                    dataDirOnHost,
 		OverrideAWSLogsExecutionRole:     overrideAWSLogsExecutionRoleEnabled,
 		CgroupPath:                       cgroupPath,
+		TaskMetadataSteadyStateRate:      steadyStateRate,
+		TaskMetadataBurstRate:            burstRate,
 	}, err
 }
 
@@ -432,6 +443,27 @@ func getTaskCPUMemLimitEnabled() Conditional {
 		}
 	}
 	return taskCPUMemLimitEnabled
+}
+
+func getTaskMetadataThrottles() (int, int) {
+	var steadyStateRate, burstRate int
+	rpsLimitEnvVal := os.Getenv("ECS_TASK_METADATA_RPS_LIMIT")
+	rpsLimitSplits := strings.Split(rpsLimitEnvVal, ",")
+	if len(rpsLimitSplits) != 2 {
+		seelog.Warn(`Invalid format for "ECS_TASK_METADATA_RPS_LIMIT", expected: "rateLimit,burst"`)
+		return 0, 0
+	}
+	steadyStateRate, err := strconv.Atoi(strings.TrimSpace(rpsLimitSplits[0]))
+	if err != nil {
+		seelog.Warnf(`Invalid format for "ECS_TASK_METADATA_RPS_LIMIT", expected integer for steady state rate: %v`, err)
+		return 0, 0
+	}
+	burstRate, err = strconv.Atoi(strings.TrimSpace(rpsLimitSplits[1]))
+	if err != nil {
+		seelog.Warnf(`Invalid format for "ECS_TASK_METADATA_RPS_LIMIT", expected integer for burst rate: %v`, err)
+		return 0, 0
+	}
+	return steadyStateRate, burstRate
 }
 
 func parseEnvVariableUint16(envVar string) uint16 {
@@ -555,6 +587,12 @@ func (cfg *Config) validateAndOverrideBounds() error {
 	if cfg.NumImagesToDeletePerCycle < minimumNumImagesToDeletePerCycle {
 		seelog.Warnf("Invalid value for number of images to delete for image cleanup, will be overridden with the default value: %d. Parsed value: %d, minimum value: %d.", DefaultImageDeletionAge, cfg.NumImagesToDeletePerCycle, minimumNumImagesToDeletePerCycle)
 		cfg.NumImagesToDeletePerCycle = DefaultNumImagesToDeletePerCycle
+	}
+
+	if cfg.TaskMetadataSteadyStateRate <= 0 || cfg.TaskMetadataBurstRate <= 0 {
+		seelog.Warnf("Invalid values for rate limits, will be overridden with default values: %d,%d.", DefaultTaskMetadataSteadyStateRate, DefaultTaskMetadataBurstRate)
+		cfg.TaskMetadataSteadyStateRate = DefaultTaskMetadataSteadyStateRate
+		cfg.TaskMetadataBurstRate = DefaultTaskMetadataBurstRate
 	}
 
 	cfg.platformOverrides()